Fixed a bug when reusing a previously deleted attribute slot.
git-svn-id: https://open-delta.googlecode.com/svn/trunk@522 dd429c29-c832-59a0-e864-46e68368aa0b

+++ delta-editor/src/main/java/au/org/ala/delta/editor/slotfile/VOItemDesc.java
@@ -374,7 +374,7 @@ public class VOItemDesc extends VOImageHolderDesc implements INameHolder {
 			}
 			if (bestOffset != -1) {
 				dataSeek(_attrOffset + bestOffset + SIZE_OF_INT_IN_BYTES);
-				_deletedAttributes.remove(bestOffset);
+				_deletedAttributes.remove((Integer)bestOffset);
 			} else { // Couldn't find a suitable free block.
 				trailerBuf = dupTrailingData(_fixedData.attribEnd);
 				if (trailerBuf != null) {
